ui: Fix implicit sign convert warn

+++ libs/ui/src/widget.c
@@ -449,7 +449,7 @@ static UiSelectFlags ui_select_dropdown(
   ui_layout_resize(canvas, anchor, ui_vector(0, lastRect.height), UiBase_Absolute, Ui_Y);
 
   for (i32 i = 0; i != (i32)entryCount; ++i) {
-    const i32 optionIndex = (dir == Ui_Up ? entryCount - 1 - i : i) - (i32)opts->allowNone;
+    const i32 optionIndex = (dir == Ui_Up ? (i32)entryCount - 1 - i : i) - (i32)opts->allowNone;
 
     const UiId     optionId     = ui_canvas_id_peek(canvas);
     const UiStatus optionStatus = ui_canvas_elem_status(canvas, optionId);
